feat: вынести rate limiter в отдельный файл
