+++ utils/limiter.js
@@ -1,0 +1,12 @@
+const rateLimit = require('express-rate-limit');
+
+const limiter = rateLimit(
+  {
+    windowMs: 15 * 60 * 1000, // 15 minutes
+    max: 100, // Limit each IP to 100 requests per `window` (here, per 15 minutes)
+    standardHeaders: true, // Return rate limit info in the `RateLimit-*` headers
+    legacyHeaders: false, // Disable the `X-RateLimit-*` headers
+  },
+);
+
+module.exports = limiter;
